updated the code to remove -fpermissive option from the unix command for compiling crhm code.

+++ crhmcode_gcc/codeconsole/CRHMmain.cpp
@@ -789,7 +789,7 @@ try {
 								TSeries * cdSeries = new TSeries(((int)(Dif * thisVar->FileData->Freq))*thisVar->FileData->ModN);
 
 								//move inside to avoid null ptr exception - Matt
-								cdSeries->Tag = (int)thisVar;
+								cdSeries->Tag = (long)thisVar;
 
 								cdSeries->Title = SS;
 							}
@@ -1969,7 +1969,7 @@ void  CRHMmain::RunClick(void) {
 
 		thisVar = (ClassVar *)(ListBox3->Objects[ii]);
 
-		cdSeries[ii]->Tag = (int)thisVar;
+		cdSeries[ii]->Tag = (long)thisVar;
 
 		string S = ListBox3->Strings[ii];
 		cdSeries[ii]->Title = S;
